Connect6 분리
