+++ src/Connect6Type.h
@@ -1,0 +1,10 @@
+#ifndef CONNECT6TYPE_H
+#define CONNECT6TYPE_H
+
+enum Piece  {EMPTY, BLACK, WHITE};
+enum Status {READY, START, BLACK1, BLACK2, WHITE1, WHITE2,
+             BLACKWIN, WHITEWIN};
+
+constexpr int BOARDSIZE = 19;
+
+#endif // CONNECT6TYPE_H
